fix: DB 관련 부분 수정

+++ src/main/java/com/kdt_proj2_be/domain/ScrapType.java
@@ -19,7 +19,7 @@ public class ScrapType {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer scrapTypeId;
+    private Long scrapTypeId;
 
     private String scrapType;
